requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
python-dotenv
cryptography
werkzeug